requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0

pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
python-jose==3.3.0
motor==3.6.0
pymongo==4.9.2
pypdf==5.3.0
numpy==2.2.2
redis==5.2.1
openai>=1.0.0,<2.0.0
python-multipart==0.0.20
bcrypt==5.0.0
email-validator==2.1.1